+++ topdown_shooter/src/player_controller.c
@@ -21,7 +21,7 @@ void tds_player_damage(SFRcomponent_t* player_controller_comp, uint32_t damage) 
 }
 
 void tds_player_controller_update(SFRcomponent_t* component, float delta_time) {
-  vec2 direction = { 0, 0 };
+  vec3 direction = { 0, 0, 0 };
   if (sfr_input_keyboard(SFR_INPUT_PRESS, SFR_KEY_A)) {
     direction[X] = -1;
   } 
@@ -34,6 +34,15 @@ void tds_player_controller_update(SFRcomponent_t* component, float delta_time) {
   if (sfr_input_keyboard(SFR_INPUT_PRESS, SFR_KEY_W)) {
     direction[Y] = 1;
   } 
+
+  if (direction[X] != 0 || direction[Y] != 0) {
+    TDSplayer_controller_t* controller = ((TDSplayer_controller_t*)component->data);
+    glm_vec3_scale(direction, controller->move_speed * delta_time, direction);
+    
+    SFRtransform_t* transform = ((SFRtransform_t*)component->owner->components[0]->data);
+    glm_vec3_add(direction, transform->position, transform->position);
+  }
+
 
   // printf("movement [%f, %f]\n", direction[X], direction[Y]);
 }